Refactor code for clarity and fixed the Serialize error with Newtonsoft.
